good bye selenium
